refactor(portabletext-editor): split util fn out in more spesific files
This will remove the need of duplicating the normalizeSelection fn in the jest collab environment
