+++ packages/@sanity/portable-text-editor/src/editor/Editable.tsx
@@ -20,7 +20,8 @@ import {PortableTextBlock} from '../types/portableText'
 import {HotkeyOptions} from '../types/options'
 import {toSlateValue, isEqualToEmptyEditor} from '../utils/values'
 import {hasEditableTarget, setFragmentData} from '../utils/copyPaste'
-import {normalizeSelection, toPortableTextRange, toSlateRange} from '../utils/selection'
+import {normalizeSelection} from '../utils/selection'
+import {toPortableTextRange, toSlateRange} from '../utils/ranges'
 import {debugWithName} from '../utils/debug'
 import {KEY_TO_SLATE_ELEMENT, KEY_TO_VALUE_ELEMENT} from '../utils/weakMaps'
 import {createWithEditableAPI} from './plugins/createWithEditableAPI'
