+++ packages/@sanity/portable-text-editor/test/setup/collaborative.jest.env.ts
@@ -2,7 +2,8 @@ import NodeEnvironment from 'jest-environment-node'
 import puppeteer from 'puppeteer'
 import ipc from 'node-ipc'
 import {isEqual} from 'lodash'
-import {EditorSelection, PortableTextBlock, EditorSelectionPoint} from '../../src'
+import {EditorSelection, PortableTextBlock} from '../../src'
+import {normalizeSelection} from '../../src/utils/selection'
 
 ipc.config.id = 'collaborative-jest-environment-ipc-client'
 ipc.config.retry = 1500
@@ -231,58 +232,4 @@ export default class CollaborationEnvironment extends NodeEnvironment {
         })
       )
   }
-}
-
-function normalizeSelection(selection: EditorSelection, value: PortableTextBlock[] | undefined) {
-  if (!selection || !value || value.length === 0) {
-    return null
-  }
-  let newAnchor: EditorSelectionPoint | null = null
-  let newFocus: EditorSelectionPoint | null = null
-  const {anchor, focus} = selection
-  if (anchor && value.find((blk) => isEqual({_key: blk._key}, anchor.path[0]))) {
-    newAnchor = normalizePoint(anchor, value)
-  }
-  if (focus && value.find((blk) => isEqual({_key: blk._key}, focus.path[0]))) {
-    newFocus = normalizePoint(focus, value)
-  }
-  if (newAnchor && newFocus) {
-    return {anchor: newAnchor, focus: newFocus}
-  }
-  return null
-}
-
-function normalizePoint(point: EditorSelectionPoint, value: PortableTextBlock[]) {
-  if (!point || !value) {
-    return null
-  }
-  const newPath: any = []
-  let newOffset: number = point.offset || 0
-  const blockKey =
-    typeof point.path[0] === 'object' && '_key' in point.path[0] && point.path[0]._key
-  const childKey =
-    typeof point.path[2] === 'object' && '_key' in point.path[2] && point.path[2]._key
-  const block: PortableTextBlock | undefined = value.find((blk) => blk._key === blockKey)
-  if (block) {
-    newPath.push({_key: block._key})
-  } else {
-    return null
-  }
-  if (block && point.path[1] === 'children') {
-    if (!block.children || block.children.length === 0) {
-      return null
-    }
-    const child = block.children.find((cld: any) => cld._key === childKey)
-    if (child) {
-      newPath.push('children')
-      newPath.push({_key: child._key})
-      newOffset =
-        child.text && child.text.length >= point.offset
-          ? point.offset
-          : (child.text && child.text.length) || 0
-    } else {
-      return null
-    }
-  }
-  return {path: newPath, offset: newOffset}
 }
